refactor: extract RadioGroupOptionPicker from flavor screen
